+++ src/aks-preview/azext_aks_preview/_help.py
@@ -491,6 +491,9 @@ helps['aks create'] = """
         - name: --nodepool-asg-ids
           type: string
           short-summary: The IDs of the application security groups to which the node pool's network interface should belong. When specified, format should be a comma-separated list of IDs.
+        - name: --node-public-ip-tags
+          type: string
+          short-summary: The ipTags of the node public IPs.
     examples:
         - name: Create a Kubernetes cluster with an existing SSH public key.
           text: az aks create -g MyResourceGroup -n MyManagedCluster --ssh-key-value /path/to/publickey
@@ -1300,6 +1303,9 @@ helps['aks nodepool add'] = """
         - name: --asg-ids
           type: string
           short-summary: The IDs of the application security groups to which the node pool's network interface should belong. When specified, format should be a comma-separated list of IDs.
+        - name: --node-public-ip-tags
+          type: string
+          short-summary: The ipTags of the node public IPs.
     examples:
         - name: Create a nodepool in an existing AKS cluster with ephemeral os enabled.
           text: az aks nodepool add -g MyResourceGroup -n nodepool1 --cluster-name MyManagedCluster --node-osdisk-type Ephemeral --node-osdisk-size 48