+++ src/aks-preview/azext_aks_preview/agentpool_decorator.py
@@ -35,6 +35,7 @@ logger = get_logger(__name__)
 AgentPool = TypeVar("AgentPool")
 AgentPoolsOperations = TypeVar("AgentPoolsOperations")
 PortRange = TypeVar("PortRange")
+IPTag = TypeVar("IPTag")
 
 
 # pylint: disable=too-few-public-methods
@@ -304,6 +305,17 @@ class AKSPreviewAgentPoolContext(AKSAgentPoolContext):
             ))
         return port_ranges
 
+    def get_ip_tags(self) -> Union[List[IPTag], None]:
+        ip_tags = self.raw_param.get("node_public_ip_tags")
+        res = []
+        if ip_tags:
+            for k, v in ip_tags.items():
+                res.append(self.models.IPTag(
+                    ip_tag_type=k,
+                    tag=v,
+                ))
+        return res
+
 
 class AKSPreviewAgentPoolAddDecorator(AKSAgentPoolAddDecorator):
     def __init__(
@@ -402,11 +414,15 @@ class AKSPreviewAgentPoolAddDecorator(AKSAgentPoolAddDecorator):
 
         asg_ids = self.context.get_asg_ids()
         allowed_host_ports = self.context.get_allowed_host_ports()
-        if asg_ids and allowed_host_ports:
-            agentpool.network_profile = self.models.AgentPoolNetworkProfile(
-                application_security_groups=asg_ids,
-                allowed_host_ports=allowed_host_ports,
-            )
+        agentpool.network_profile = self.models.AgentPoolNetworkProfile()
+        if allowed_host_ports is not None:
+            agentpool.network_profile.allowed_host_ports = allowed_host_ports
+            agentpool.network_profile.application_security_groups = asg_ids
+
+        ip_tags = self.context.get_ip_tags()
+        if ip_tags:
+            agentpool.network_profile.node_public_ip_tags = ip_tags
+
         return agentpool
 
     def construct_agentpool_profile_preview(self) -> AgentPool:
@@ -490,7 +506,7 @@ class AKSPreviewAgentPoolUpdateDecorator(AKSAgentPoolUpdateDecorator):
 
         asg_ids = self.context.get_asg_ids()
         allowed_host_ports = self.context.get_allowed_host_ports()
-        if asg_ids or allowed_host_ports:
+        if not agentpool.network_profile and (asg_ids or allowed_host_ports):
             agentpool.network_profile = self.models.AgentPoolNetworkProfile()
         if asg_ids is not None:
             agentpool.network_profile.application_security_groups = asg_ids