+++ src/aks-preview/HISTORY.rst
@@ -12,10 +12,27 @@ To release a new version, please select a new version number (usually plus 1 to 
 Pending
 +++++++
 
+0.5.115
++++++++
+
+* Support node public IPTags by `az aks create` and `az aks nodepool add`.
+
+0.5.114
++++++++
+
+* Fix `az aks create` and `az aks nodepool add` commands failing on adding nodepool with managed ApplicationSecurityGroups.
+
+0.5.113
++++++++
+
+* Fix workload identity update error after oidc issure GA in azure-cli.
+* Fix `az aks update` command failing on SP-based cluster blocked by validation in AzureMonitorMetrics Addon, see issue `\#5336 <https://github.com/Azure/azure-cli-extensions/issues/5488>`_.
+* Fix `az aks update` command failing on changes not related to outbound type conversion, see issue `\#24430 https://github.com/Azure/azure-cli/issues/24430>`_.
+
 0.5.112
 +++++++
 
-* Add --outbound-type to update managed cluster command.
+* Add `--outbound-type` to update managed cluster command.
 
 0.5.111
 +++++++